+++ EmployeeReimbursementSystem/RepositoryLayer/TicketRepository.cs
@@ -7,15 +7,9 @@ using Microsoft.Data.SqlClient;
 
 using ModelLayer;
 
-/**
- * TODO, Reduce the number of repeated lines of code and increase readability
- * - Identify where code keeps getting reused, seperate functionality into 
- *   seperate methods.
- * - Add in loggers.
- */
 namespace RepositoryLayer;
 public interface ITicketRepository {
-    ReimburseTicket PostTicket(string guid, string r, double a, string d, DateTime t, int eId);
+    ReimburseTicket PostTicket(string guid, string r, int a, string d, DateTime t, int eId);
     ReimburseTicket GetTicket(string ticketId);
     ReimburseTicket UpdateTicket(string ticketId, int statusId);
     List<ReimburseTicket> GetTickets(int employeeId);
@@ -54,7 +48,7 @@ public class TicketRepository : ITicketRepository {
         }
     }
 
-    public ReimburseTicket PostTicket(string guid, string r, double a, string d, DateTime t, int eId) {
+    public ReimburseTicket PostTicket(string guid, string r, int a, string d, DateTime t, int eId) {
         string conString = File.ReadAllText("../../ConString.txt");
         using(SqlConnection connection = new SqlConnection(conString)) {
             string insertTicketQuery = "INSERT INTO Ticket (TicketId, Reason, Amount, Description, StatusId, RequestDate, EmployeeId) VALUES (@guid, @r, @a, @d, 0, @t, @eId);";
@@ -104,7 +98,7 @@ public class TicketRepository : ITicketRepository {
                         return new ReimburseTicket(
                             (string) reader[0],
                             (string) reader[1],
-                            (double) reader[2],
+                            (int) reader[2],
                             (string) reader[3],
                             (int) reader[4],
                             (DateTime) reader[5],
@@ -140,7 +134,7 @@ public class TicketRepository : ITicketRepository {
                         ReimburseTicket newTicket = new ReimburseTicket(
                             (string) reader[0],
                             (string) reader[1],
-                            (double) reader[2],
+                            (int) reader[2],
                             (string) reader[3],
                             (int) reader[4],
                             (DateTime) reader[5],
@@ -181,7 +175,7 @@ public class TicketRepository : ITicketRepository {
                             ReimburseTicket newTicket = new ReimburseTicket(
                                 (string) reader[0],
                                 (string) reader[1],
-                                (double) reader[2],
+                                (int) reader[2],
                                 (string) reader[3],
                                 (int) reader[4],
                                 (DateTime) reader[5],
@@ -222,7 +216,7 @@ public class TicketRepository : ITicketRepository {
                             ReimburseTicket newTicket = new ReimburseTicket(
                                 (string) reader[0],
                                 (string) reader[1],
-                                (double) reader[2],
+                                (int) reader[2],
                                 (string) reader[3],
                                 (int) reader[4],
                                 (DateTime) reader[5],
